extract useWormBaseTableFilter
